pyproject.toml
[tool.poetry]
name = "metadmg-simulations"
version = "0.1.0"
description = ""
authors = ["Christian Michelsen <christianmichelsen@gmail.com>"]
readme = "README.md"
# packages = [{include = "metadmg_simulations"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
pysam = "^0.19.1"
parse = "^1.19.0"
pandas = "^1.4.3"
seaborn = "^0.11.2"
biopython = "^1.79"
pyarrow = "^9.0.0"
dash = {version = "^2.6.1", extras = ["diskcache"]}
metadmg = {version = "^0.28.4", extras = ["fit"]}
numpyro = "^0.10.0"
llvmlite = "^0.39.0"
numpy = "1.22"
dash-bootstrap-components = "^1.2.1"
cycler = "^0.11.0"
Pillow = "^9.2.0"
kiwisolver = "^1.4.4"
ipywidgets = "^7.7.1"
fonttools = "^4.34.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.15.0"
# black = {version = "^22.6.0", allow-prereleases = true}
black = {version = "^22.6.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
